Extract Introduction from Generator
